rev.1

+++ go/handlers.go
@@ -1,22 +1,73 @@
 package main
 
 import (
+  "net/url"
   "time"
 
+  jwt "github.com/dgrijalva/jwt-go"
   "github.com/gin-gonic/gin"
   "github.com/jinzhu/gorm"
 )
 
-type jadwals struct{
-  gorm.Model
-  Kegiatan  string
-  Jam       string
+func logoutHandlers() gin.HandlerFunc {
+  return func(c *gin.Context) {
+    c.SetCookie("token", "", 0, "/", "", false, false)
+    c.JSON(200, gin.H{"response":true})
+  }
+}
+
+func loginHandlers(cfg *Cfg) gin.HandlerFunc {
+  return func(c *gin.Context) {
+    db := cfg.DBInstance
+
+    var (
+      user []logins
+      data, _ = url.ParseQuery(c.PostForm("data"))
+    )
+
+    err := db.Where("username = ? AND password = ?", data["username"], data["password"]).Find(&user).Error
+    if err != nil {
+      c.JSON(500, gin.H{
+        "msg":err.Error(),
+        "response":false,
+      })
+      return
+    }
+
+    if len(user) > 0 {
+      var claimsData = struct {
+        jwt.StandardClaims
+        UserData gin.H
+      }{
+        StandardClaims : jwt.StandardClaims{
+          Issuer: "lumoshive",
+          ExpiresAt: time.Now().Add(time.Hour * 24).Unix(),
+        },
+        UserData: gin.H{"id":user[0].ID, "name":user[0].Name},
+      }
+
+      token := jwt.NewWithClaims(jwt.SigningMethodHS256, claimsData)
+      ss, _ := token.SignedString([]byte(cfg.SignatureKey))
+
+      c.SetCookie("token", ss, 3600, "/", "", false, true)
+      c.SetCookie("login", "true", 3600, "/", "", false, false)
+
+      c.JSON(200, gin.H{
+        "msg": "login success",
+        "response":true,
+      })
+    } else {
+      c.JSON(200, gin.H{
+        "msg":"user not found",
+        "response":false,
+      })
+    }
+  }
 }
 
 func actionHandlers(cfg *Cfg) gin.HandlerFunc {
   return func(c *gin.Context) {
     db := cfg.DBInstance
-    db.AutoMigrate(&jadwals{})
 
     var methodNotFound = func() {
       c.JSON(400, gin.H{
@@ -36,19 +87,22 @@ func actionHandlers(cfg *Cfg) gin.HandlerFunc {
       case "read"  : read(c, db)
       case "update": update(c, db)
       case "delete": delete(c, db)
-      default:
-        methodNotFound()
+      default: methodNotFound()
     }
   }
 }
 
 func create(c *gin.Context, db *gorm.DB) {
-  var data = jadwals{
-    Kegiatan: c.PostForm("Kegiatan"),
-    Jam: c.PostForm("Jam"),
-  }
-  db.Create(&data)
+  var (
+    UserData  = c.MustGet("UserData").(map[string]interface{})
+    data      = jadwals{
+                Kegiatan: c.PostForm("Kegiatan"),
+                Jam:      c.PostForm("Jam"),
+                Userid:   UserData["id"],
+              }
+  )
 
+  db.Create(&data)
   c.JSON(200, gin.H{
     "response": true,
     "msg":   "berhasil insert data",
@@ -56,9 +110,12 @@ func create(c *gin.Context, db *gorm.DB) {
 }
 
 func read(c *gin.Context, db *gorm.DB) {
-  var data []jadwals
-  db.Find(&data)
+  var (
+    data []jadwals
+    UserData = c.MustGet("UserData").(map[string]interface{})
+  )
 
+  db.Where("userid = ?", UserData["id"]).Find(&data)
   c.JSON(200, gin.H{
     "response": true,
     "data": data,
@@ -66,14 +123,17 @@ func read(c *gin.Context, db *gorm.DB) {
 }
 
 func update(c *gin.Context, db *gorm.DB) {
-  id       := c.PostForm("id")
-  kegiatan := c.PostForm("Kegiatan")
-  jam      := c.PostForm("Jam")
+  var (
+    data []jadwals
+    UserData = c.MustGet("UserData").(map[string]interface{})
+    id       = c.PostForm("id")
+    kegiatan = c.PostForm("Kegiatan")
+    jam      = c.PostForm("Jam")
+  )
 
-  var data jadwals
+  db.Where("id=? AND userid=?", id, UserData["id"]).Find(&data)
 
-  err := db.First(&data, id).Error
-  if err != nil {
+  if len(data) < 1 {
     c.JSON(200, gin.H{
       "response": false,
       "msg": "data not found",
@@ -81,7 +141,7 @@ func update(c *gin.Context, db *gorm.DB) {
     return
   }
 
-  err = db.Exec("UPDATE jadwals SET kegiatan=?, jam=?, updated_at=? WHERE id=?", kegiatan, jam, time.Now(), id).Error
+  err := db.Exec("UPDATE jadwals SET kegiatan=?, jam=?, updated_at=? WHERE id=? AND userid=?", kegiatan, jam, time.Now(), id, UserData["id"]).Error
 
   if err != nil {
     c.JSON(200, gin.H{
@@ -97,8 +157,13 @@ func update(c *gin.Context, db *gorm.DB) {
 }
 
 func delete(c *gin.Context, db *gorm.DB) {
-  id := c.PostForm("id")
-  err := db.Exec("DELETE FROM jadwals WHERE id=?", id).Error
+  var (
+    id       = c.PostForm("id")
+    UserData = c.MustGet("UserData").(map[string]interface{})
+  )
+
+  err := db.Exec("DELETE FROM jadwals WHERE id=? AND userid=?", id, UserData["id"]).Error
+
   if err != nil {
     c.JSON(200, gin.H{
       "response": false,
